+++ docs/changelog.rst
@@ -6,8 +6,8 @@ Changelog
 * Added support for multiple links between nodes (Parallel_Link_Model)
 * Cached parallel_lsp_groups in Model and Parallel_Link_Model objects (performance optimization)
 * Added check for multiple links between nodes in Model object (not allowed)
-* Added Parent Class MasterModel to hold common defs for Model and Parallel_Link_Model subclasses
-* Added simulation_diagnostics def in MasterModel that gives potentially useful diagnostic info about the simulation results
+* Added Parent Class _MasterModel to hold common defs for Model and Parallel_Link_Model subclasses
+* Added simulation_diagnostics def in _MasterModel that gives potentially useful diagnostic info about the simulation results
 * Simple user interface (beta feature) supports RSVP LSPs
 
 1.5